feat(workstation): adapt setup

+++ init.sh
@@ -13,20 +13,36 @@ check_command() {
   fi
 }
 
-# install yay if not present
-if ! check_command yay; then
-  # clean up existing directory if exists
-  [ -d "/tmp/yay" ] && rm -rf /tmp/yay
-  git clone https://aur.archlinux.org/yay.git /tmp/yay
-  cd /tmp/yay
-  yes | makepkg -si
-  rm -rf /tmp/yay
+# check if pacman exists
+is_arch() {
+  command -v pacman &> /dev/null
+}
+
+# install yay if on arch and not present
+if is_arch; then
+  if ! check_command yay; then
+    # clean up existing directory if exists
+    [ -d "/tmp/yay" ] && rm -rf /tmp/yay
+    git clone https://aur.archlinux.org/yay.git /tmp/yay
+    cd /tmp/yay
+    yes | makepkg -si
+    rm -rf /tmp/yay
+  fi
 fi
 
 # install mise if not present
 if ! check_command mise; then
-  yay -Syu --needed --noconfirm mise
-  eval "$(mise activate bash)"
-else
-  eval "$(mise activate bash)"
+  if is_arch; then
+    yay -Syu --needed --noconfirm mise
+  else
+    curl https://mise.run | sh
+  fi
+fi
+
+# activate mise
+mise activate --shims
+
+if [ -n "${WEB_HOST}" ]; then
+  echo "WEB_HOST is set. running workstation setup..."
+  task install:workstation
 fi
